+++ route/src/index.rs
@@ -1,5 +1,4 @@
 use maud::{html, Markup};
-use service::*;
 
 #[get("/")]
 fn index() -> Markup {
