+++ route/src/lib.rs
@@ -9,7 +9,7 @@ extern crate data;
 extern crate service;
 extern crate guard;
 extern crate transfer;
-#[macro_use] extern crate serde_derive;
+extern crate serde_derive;
 extern crate serde;
 extern crate serde_json;
